- background and foreground selection colors setup moved to form files - LuaFileFilter moved from method in PriceEditorDialogPresenter to the separate class - added error message on Tamriel Trade Center price table file read error - naming changes - small fixes
